Use tailwind classes to add styling to page.tsx
I told Claude to "Use tailwind classes to add styling to page.tsx" and it added several tailwind classes to improve the appearance of the page. I may go through and tweak these but they look pretty good for now and design is not my forte. If given any kind of reference like a Figma design or even just looking at the existing style already in place in the app, I could easily nail the desired look, but I'm not good at coming up with good design from scratch.