Require DATABASE_URL
My IDE was complaining about "insert" not existing on the db object because the mock that is returned if DATABASE_URL is not set does not have an insert method. Since this kind of mocking wouldn't be used in a production-ready app, I decided to just remove that mock and instead throw an error if DATABASE_URL is not set. This way the db object is always a proper Drizzle ORM database instance and has all the expected methods.